requirements.txt
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
thop>=0.1.1
torch==1.13.1
torchvision==0.14.1
tqdm>=4.66.3
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0